container/1.featurize/requirements.txt
scikit-learn==1.0.2
google-cloud-storage==2.1.0; python_version > '3.6'